fix: adapter for production company & country not recognized as .ts files
