+++ src/app/adapters/production-country.adapter.ts
@@ -2,20 +2,18 @@ import {Injectable} from "@angular/core";
 
 import {Adapter} from "./base.adapter";
 
-import {ProductionCompany} from "../models/movie.model";
-import {TmdbProductionCompany} from "../models/tmdb/tmdb-movie.model";
+import {ProductionCountry} from "../models/movie.model";
+import {TmdbProductionCountry} from "../models/tmdb/tmdb-movie.model";
 
 
 @Injectable({
   providedIn: "root",
 })
-export class ProductionCompanyAdapter implements Adapter<ProductionCompany> {
-  adapt(item: TmdbProductionCompany): ProductionCompany {
-    return new ProductionCompany(
-      item.id,
+export class ProductionCountryAdapter implements Adapter<ProductionCountry> {
+  adapt(item: TmdbProductionCountry): ProductionCountry {
+    return new ProductionCountry(
       item.name,
-      item.origin_country,
-      item.logo_path,
+      item.iso_3166_1,
     );
   }
 }
